subcategorizeAll
